requirements.txt
absl-py==2.0.0
apturl==0.5.2
astroid==2.9.3
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
bnlearn==0.8.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
colourmap==1.1.16
comm==0.1.4
command-not-found==0.3
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
d3blocks==1.4.7
d3graph==2.5.0
datazets==0.1.9
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
df2onehot==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
elasticgraph==0.2.0
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.14.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.10.0
funcsigs==1.0.2
future==0.18.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
grpcio-tools==1.62.1
gyp==0.1
h5py==3.10.0
hashport==1.2303.3
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
ismember==1.0.2
isort==5.6.4
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
keras==2.14.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
libclang==16.0.6
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5
MarkupSafe==2.0.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mininet @ file:///home/rishav/Desktop/mininet
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mplcyberpunk==0.7.0
mpmath==1.3.0
mypy-extensions==0.4.3
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
ovs==2.17.9
packaging==23.2
pandas==1.5.3
paramiko==2.9.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pgmpy==0.1.24
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pyflakes==2.4.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypickle==1.1.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-louvain==0.16
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scapy==2.5.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
SecretStorage==3.3.1
six==1.16.0
sortedcontainers==2.1.0
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ucimlrepo==0.0.3
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.8
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
